+++ apps/web/components/WishlistButton/types.ts
@@ -1,0 +1,6 @@
+import { Product } from '@plentymarkets/shop-api';
+
+export type WishlistButtonProps = {
+  product: Product;
+  quantity?: number;
+};
