+++ apps/web/components/ui/PurchaseCard/PurchaseCard.vue
@@ -3,9 +3,21 @@
     class="p-4 xl:p-6 md:border md:border-neutral-100 md:shadow-lg md:rounded-md md:sticky md:top-40"
     data-testid="purchase-card"
   >
-    <h1 class="mb-1 font-bold typography-headline-4" data-testid="product-name">
-      {{ productName }}
-    </h1>
+    <div class="flex justify-between mt-4 gap-x-4">
+      <h1 class="mb-1 font-bold typography-headline-4" data-testid="product-name">
+        {{ productGetters.getName(product) }}
+      </h1>
+      <WishlistButton v-if="isDesktop" :product="product" :quantity="quantitySelectorValue">
+        {{ t('addProductToWishlist') }}
+      </WishlistButton>
+      <WishlistButton
+        v-else
+        square
+        class="bottom-0 right-0 mr-2 mb-2 bg-white ring-1 ring-inset ring-neutral-200 !rounded-full"
+        :product="product"
+        :quantity="quantitySelectorValue"
+      />
+    </div>
     <Price
       :price="currentActualPrice"
       :normal-price="normalPrice"
@@ -68,22 +80,7 @@
           </SfButton>
         </SfTooltip>
       </div>
-      <div class="flex justify-end mt-4 gap-x-4">
-        <SfButton
-          variant="tertiary"
-          size="sm"
-          :aria-label="t('addProductToWishlist', productName)"
-          :disabled="wishlistLoading"
-          @click="onWishlistClick()"
-        >
-          <SfLoaderCircular v-if="wishlistLoading" class="flex justify-center items-center" size="sm" />
-          <template v-else>
-            <SfIconFavoriteFilled v-if="isWishlistItem(variationId)" size="sm" />
-            <SfIconFavorite v-else size="sm" />
-            {{ t('addProductToWishlist') }}
-          </template>
-        </SfButton>
-      </div>
+
       <div class="mt-4 typography-text-xs flex gap-1">
         <span>{{ $t('asterisk') }}</span>
         <span v-if="showNetPrices">{{ $t('itemExclVAT') }}</span>
@@ -120,8 +117,6 @@ import {
   SfIconShoppingCart,
   SfLoaderCircular,
   SfTooltip,
-  SfIconFavorite,
-  SfIconFavoriteFilled,
 } from '@storefront-ui/vue';
 import type { PurchaseCardProps } from '~/components/ui/PurchaseCard/types';
 
@@ -132,15 +127,14 @@ const props = defineProps<PurchaseCardProps>();
 
 const { product } = toRefs(props);
 
+const { isDesktop } = useBreakpoints();
 const { getPropertiesForCart } = useProductOrderProperties();
 const { send } = useNotification();
 const { addToCart, loading } = useCart();
 const { getPropertiesPrice } = useProductOrderProperties();
 const { t } = useI18n();
-const { isWishlistItem, interactWithWishlist } = useWishlist();
 
 const quantitySelectorValue = ref(1);
-const wishlistLoading = ref(false);
 
 const currentActualPrice = computed(
   () =>
@@ -202,15 +196,6 @@ const scrollToReviewsAccordion = () => {
 };
 
 const isSalableText = computed(() => (productGetters.isSalable(product.value) ? '' : t('itemNotAvailable')));
-
-const productName = computed(() => productGetters.getName(product.value));
-const variationId = computed(() => productGetters.getVariationId(product.value));
-
-const onWishlistClick = async () => {
-  wishlistLoading.value = true;
-  await interactWithWishlist(variationId.value, quantitySelectorValue.value);
-  wishlistLoading.value = false;
-};
 
 const scrollToReviews = () => {
   if (!isReviewsAccordionOpen()) {
